Split out bundler methods into SimulatedBundler
