+++ inpage/src/bundlers/IBundler.ts
@@ -1,0 +1,11 @@
+import EthereumRpc from '../EthereumRpc';
+
+type IBundler = Pick<
+  EthereumRpc.Handlers,
+  | 'eth_sendUserOperation'
+  | 'eth_estimateUserOperationGas'
+  | 'eth_getUserOperationReceipt'
+  | 'eth_supportedEntryPoints'
+>;
+
+export default IBundler;
