+++ inpage/src/waxPrivate.ts
@@ -1,0 +1,7 @@
+/**
+ * This symbol is not exported to users of the library so that functions that
+ * should be internal to wax can test against it.
+ */
+const waxPrivate = Symbol('waxPrivate');
+
+export default waxPrivate;
